netkeiba_crawler_gcf/requirements.txt
functions_framework==3.4.0
itemadapter==0.8.0
pandas==1.3.4
protobuf==4.23.4
Requests==2.31.0
Scrapy==2.9.0

google-cloud-bigquery~=3.11.4
db-dtypes~=1.1.1